refactor : 랜덤으로 뽑은 로또 검증을 위한 함수도 검증하기 위해서 Util로 이동
